// src/app/models/indicadores.ts
export class Indicadores {
    id: number;
    start: number;
    aire: number;
    enraizamiento: number;
    vegetativo: number;
    floracion: number;
    calefactor: number;
    ventilador: number;
    alertaTemperatura: number;
    humificador: number;
    deshumificador: number;
    alertaHumedad: number;
    bomba: number;
    alertaNivel: number;
    phElevado: number;
    phBajo: number;
    alertaPh: number;
}